move all tests for the serviceability smart contract into a single folder.
Add comprehensive tests for multicast group and user allowlist functionalities

- Implement tests for multicast group publisher allowlist management, including adding and removing entries.
- Create tests for multicast group subscriber allowlist management, verifying the addition and removal of subscribers.
- Develop tests for multicast group lifecycle, covering creation, activation, suspension, reactivation, updating, and deletion.
- Introduce user allowlist tests to validate the addition and removal of users from the global allowlist.
- Enhance user management tests, including creation, activation, suspension, resumption, updating, and deletion of user accounts.
- Utilize helper functions for transaction execution and account data retrieval to streamline test processes.

- Remove trailing whitespace from device and link module files
