+++ smartcontract/programs/doublezero-serviceability/src/lib.rs
@@ -14,5 +14,4 @@ pub mod processors;
 pub mod programversion;
 pub mod seeds;
 pub mod state;
-pub mod tests;
 pub mod types;
