+++ smartcontract/programs/doublezero-serviceability/src/processors/contributor/mod.rs
@@ -2,5 +2,4 @@ pub mod create;
 pub mod delete;
 pub mod resume;
 pub mod suspend;
-pub mod test;
 pub mod update;
