+++ smartcontract/programs/doublezero-serviceability/src/processors/device/mod.rs
@@ -6,5 +6,3 @@ pub mod reject;
 pub mod resume;
 pub mod suspend;
 pub mod update;
-
-pub mod test;
