+++ smartcontract/programs/doublezero-serviceability/src/processors/location/mod.rs
@@ -3,5 +3,3 @@ pub mod delete;
 pub mod resume;
 pub mod suspend;
 pub mod update;
-
-pub mod test;
