+++ smartcontract/programs/doublezero-serviceability/src/processors/multicastgroup/allowlist/publisher/mod.rs
@@ -1,3 +1,2 @@
 pub mod add;
 pub mod remove;
-pub mod test;
