+++ smartcontract/programs/doublezero-serviceability/src/processors/multicastgroup/mod.rs
@@ -7,5 +7,4 @@ pub mod reactivate;
 pub mod reject;
 pub mod subscribe;
 pub mod suspend;
-pub mod test;
 pub mod update;
